Fix pre-existing CI build failures across multiple files
- src/CMakeLists.txt: Link libcurl for non-Windows (auto_updater uses curl)
- src/common/gui/error_message.h: Suppress false positive
  performance-unnecessary-value-param for pass-by-value+move pattern
- src/gui-client/connection_widget.cpp: Fix bugprone-integer-division by
  using qreal for shield drawing coordinates; fix readability-implicit-
  bool-conversion by using explicit nullptr comparisons
- src/gui-client/ipc_client_manager.cpp: Fix readability-implicit-bool-
  conversion for std::get_if pointer checks
- tests/unit/CMakeLists.txt: Add Qt6::Test component for QSignalSpy
- tests/unit/connection_widget_tests.cpp: Fix implicit widening
  conversion in updateMetrics calls

+++ tests/unit/connection_widget_tests.cpp
@@ -113,7 +113,7 @@ TEST_F(ConnectionWidgetTest, StateTransitionToError) {
 
 TEST_F(ConnectionWidgetTest, UpdateMetricsBasic) {
   widget_->setConnectionState(ConnectionState::kConnected);
-  widget_->updateMetrics(50, 1024 * 100, 1024 * 50);  // 50ms, 100KB/s, 50KB/s
+  widget_->updateMetrics(50, 1024U * 100U, 1024U * 50U);  // 50ms, 100KB/s, 50KB/s
   // Should not crash
 }
 
@@ -355,7 +355,7 @@ TEST_F(ConnectionWidgetTest, RapidStateChanges) {
 TEST_F(ConnectionWidgetTest, UpdateMetricsRapidly) {
   widget_->setConnectionState(ConnectionState::kConnected);
   for (int i = 0; i < 100; ++i) {
-    widget_->updateMetrics(i % 100, i * 1024, i * 512);
+    widget_->updateMetrics(i % 100, static_cast<uint64_t>(i) * 1024U, static_cast<uint64_t>(i) * 512U);
   }
 }
